main.py
#!/usr/bin/env python3
"""
KME (Key Management Entity) - Main Application Entry Point

Version: 1.0.0
Author: KME Development Team
Description: ETSI GS QKD 014 V1.1.1 compliant Key Management Entity
License: [To be determined]

ToDo List:
- [ ] Implement FastAPI application setup
- [ ] Add middleware for authentication
- [ ] Configure CORS settings
- [ ] Set up API routing
- [ ] Add health check endpoints
- [ ] Implement logging configuration
- [ ] Add error handling middleware
- [ ] Configure TLS settings
- [ ] Add metrics collection
- [ ] Implement graceful shutdown

Progress: 0% (Not started)
"""

import datetime
import os
import sys
from pathlib import Path

# Add the project root to Python path
project_root = Path(__file__).parent
sys.path.insert(0, str(project_root))

# Load environment variables
from dotenv import load_dotenv

load_dotenv()

import structlog
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from fastapi.middleware.trustedhost import TrustedHostMiddleware

# Import health monitoring
from app.core.health import check_health, get_health_summary

# Import performance monitoring
from app.core.performance import get_performance_monitor

# Import KME modules (to be implemented)
# from app.core.config import settings
# from app.core.logging import setup_logging
# from app.api.routes import api_router
# from app.core.middleware import AuthMiddleware



# Initialize structured logging
structlog.configure(
    processors=[
        structlog.stdlib.filter_by_level,
        structlog.stdlib.add_logger_name,
        structlog.stdlib.add_log_level,
        structlog.stdlib.PositionalArgumentsFormatter(),
        structlog.processors.TimeStamper(fmt="iso"),
        structlog.processors.StackInfoRenderer(),
        structlog.processors.format_exc_info,
        structlog.processors.UnicodeDecoder(),
        structlog.processors.JSONRenderer(),
    ],
    context_class=dict,
    logger_factory=structlog.stdlib.LoggerFactory(),
    wrapper_class=structlog.stdlib.BoundLogger,
    cache_logger_on_first_use=True,
)

logger = structlog.get_logger()

# Create FastAPI application
app = FastAPI(
    title="KME - Key Management Entity",
    description="ETSI GS QKD 014 V1.1.1 compliant Key Management Entity",
    version="1.0.0",
    docs_url="/docs" if os.getenv("DEBUG", "false").lower() == "true" else None,
    redoc_url="/redoc" if os.getenv("DEBUG", "false").lower() == "true" else None,
)

# Configure CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=os.getenv("ALLOWED_ORIGINS", "[]").split(","),
    allow_credentials=True,
    allow_methods=["GET", "POST"],
    allow_headers=["*"],
)

# Configure trusted hosts
app.add_middleware(
    TrustedHostMiddleware, allowed_hosts=os.getenv("ALLOWED_HOSTS", "*").split(",")
)

# Add custom middleware (to be implemented)
# app.add_middleware(AuthMiddleware)

# Include API routes (to be implemented)
# app.include_router(api_router, prefix="/api/v1")


@app.on_event("startup")
async def startup_event():
    """Application startup event handler"""
    logger.info("KME application starting up")

    # Initialize components (to be implemented)
    # await setup_logging()
    # await initialize_database()
    # await initialize_redis()
    # await setup_security()

    logger.info("KME application startup complete")


@app.on_event("shutdown")
async def shutdown_event():
    """Application shutdown event handler"""
    logger.info("KME application shutting down")

    # Cleanup components (to be implemented)
    # await cleanup_database()
    # await cleanup_redis()

    logger.info("KME application shutdown complete")


@app.get("/")
async def root():
    """Root endpoint"""
    return {
        "message": "KME - Key Management Entity",
        "version": "1.0.0",
        "status": "operational",
        "specification": "ETSI GS QKD 014 V1.1.1",
    }


@app.get("/health")
async def health_check():
    """Health check endpoint"""
    return await check_health()


@app.get("/health/summary")
async def health_summary():
    """Health summary endpoint"""
    return await get_health_summary()


@app.get("/health/ready")
async def health_ready():
    """Readiness check endpoint"""
    health_data = await check_health()
    if health_data["status"] in ["healthy", "degraded"]:
        return {"status": "ready", "message": "KME is ready to serve requests"}
    else:
        return {"status": "not_ready", "message": "KME is not ready to serve requests"}


@app.get("/health/live")
async def health_live():
    """Liveness check endpoint"""
    return {"status": "alive", "message": "KME is alive and responding"}


@app.get("/metrics/performance")
async def get_performance_metrics():
    """Get performance metrics"""
    monitor = get_performance_monitor()
    return {
        "api_performance": monitor.get_api_performance_summary(),
        "key_performance": monitor.get_key_performance_summary(),
        "system_performance": monitor.get_system_performance_metrics(),
        "timestamp": datetime.datetime.utcnow().isoformat(),
    }


@app.get("/metrics/api")
async def get_api_metrics():
    """Get API performance metrics"""
    monitor = get_performance_monitor()
    return {
        "api_performance": monitor.get_api_performance_summary(),
        "timestamp": datetime.datetime.utcnow().isoformat(),
    }


@app.get("/metrics/keys")
async def get_key_metrics():
    """Get key management performance metrics"""
    monitor = get_performance_monitor()
    return {
        "key_performance": monitor.get_key_performance_summary(),
        "timestamp": datetime.datetime.utcnow().isoformat(),
    }


@app.get("/metrics/system")
async def get_system_metrics():
    """Get system performance metrics"""
    monitor = get_performance_monitor()
    return {
        "system_performance": monitor.get_system_performance_metrics(),
        "timestamp": datetime.datetime.utcnow().isoformat(),
    }


@app.get("/metrics/database")
async def get_database_metrics():
    """Get database performance metrics"""
    monitor = get_performance_monitor()
    return {
        "database_performance": monitor.get_database_performance_summary(),
        "timestamp": datetime.datetime.utcnow().isoformat(),
    }
    # Health check logic (to be implemented)
    return {
        "status": "healthy",
        "timestamp": "2024-01-01T00:00:00Z",
        "version": "1.0.0",
    }


@app.get("/api/v1/keys/{slave_sae_id}/status")
async def get_status(slave_sae_id: str):
    """Get Status endpoint - Placeholder implementation"""
    # TODO: Implement actual status endpoint
    logger.info(f"Status request for slave SAE: {slave_sae_id}")

    # Record performance metric
    monitor = get_performance_monitor()
    monitor.record_api_metric("/api/v1/keys/{slave_sae_id}/status", 0.0, 200)

    # Placeholder response
    return {
        "source_KME_ID": os.getenv("KME_ID", "AAAABBBBCCCCDDDD"),
        "target_KME_ID": "EEEEFFFFGGGGHHHH",
        "master_SAE_ID": "IIIIJJJJKKKKLLLL",
        "slave_SAE_ID": slave_sae_id,
        "key_size": int(os.getenv("DEFAULT_KEY_SIZE", "352")),
        "stored_key_count": 0,
        "max_key_count": 100000,
        "max_key_per_request": int(os.getenv("MAX_KEYS_PER_REQUEST", "128")),
        "max_key_size": int(os.getenv("MAX_KEY_SIZE", "1024")),
        "min_key_size": int(os.getenv("MIN_KEY_SIZE", "64")),
        "max_SAE_ID_count": int(os.getenv("MAX_SAE_ID_COUNT", "10")),
    }


@app.post("/api/v1/keys/{slave_sae_id}/enc_keys")
async def get_key(slave_sae_id: str):
    """Get Key endpoint - Placeholder implementation"""
    # TODO: Implement actual key endpoint
    logger.info(f"Key request for slave SAE: {slave_sae_id}")

    # Record performance metric
    monitor = get_performance_monitor()
    monitor.record_api_metric("/api/v1/keys/{slave_sae_id}/enc_keys", 0.0, 200)
    monitor.record_key_metric("key_retrieval", 0.0, 1, 352)

    # Placeholder response
    return {
        "keys": [
            {
                "key_ID": "550e8400-e29b-41d4-a716-446655440000",
                "key": "placeholder-key-data-base64-encoded",
            }
        ]
    }


@app.post("/api/v1/keys/{master_sae_id}/dec_keys")
async def get_key_with_ids(master_sae_id: str):
    """Get Key with Key IDs endpoint - Placeholder implementation"""
    # TODO: Implement actual key with IDs endpoint
    logger.info(f"Key with IDs request for master SAE: {master_sae_id}")

    # Record performance metric
    monitor = get_performance_monitor()
    monitor.record_api_metric("/api/v1/keys/{master_sae_id}/dec_keys", 0.0, 200)
    monitor.record_key_metric("key_retrieval_with_ids", 0.0, 1, 352)

    # Placeholder response
    return {
        "keys": [
            {
                "key_ID": "550e8400-e29b-41d4-a716-446655440000",
                "key": "placeholder-key-data-base64-encoded",
            }
        ]
    }


if __name__ == "__main__":
    import uvicorn

    # Get configuration from environment
    host = os.getenv("KME_HOSTNAME", "localhost")
    port = int(os.getenv("KME_PORT", "8443"))
    debug = os.getenv("DEBUG", "false").lower() == "true"
    reload = os.getenv("RELOAD", "false").lower() == "true"

    logger.info(f"Starting KME server on {host}:{port}")

    uvicorn.run(
        "main:app",
        host=host,
        port=port,
        debug=debug,
        reload=reload,
        log_config=None,  # Use structlog configuration
    )
